:art::rotating_light: Separate specs into controller/action_spec
