+++ spec/requests/employees/create_spec.rb
@@ -1,0 +1,82 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe 'Employees POST /employees', type: :request, js: true do
+  # WARNING: it rarely generates duplicated username by Faker::Internet.username
+  let!(:administor) { create(:administor) }
+  let!(:employees) { create_list(:employee, 5) }
+  let(:employee) { employees.first }
+  let(:employee_id) { employee.id }
+
+  # all actions are requied logged in
+  before { post login_path, params: { username: administor.username, password: administor.password } }
+
+  context 'when request is valid' do
+    let(:valid_params) do
+      {
+        employee: {
+          fullname: 'Dr. Blah',
+          username: 'blah',
+          password: 'passworld',
+          password_confirmation: 'passworld'
+        }
+      }
+    end
+    before { post employees_path, params: valid_params }
+
+    it 'creates new employee' do
+      expect(Employee.last).to eq(assigns[:employee])
+    end
+
+    it { should redirect_to(employee_path(Employee.last.id)) }
+  end
+
+  context 'when request is invalid' do
+    let(:invalid_params) do
+      {
+        employee: {
+          fullname: '',
+          username: '',
+          password: '',
+          password_confirmation: ''
+        }
+      }
+    end
+    before { post employees_path, params: invalid_params }
+
+    it { should render_template('new') }
+  end
+
+  # duplicates on model spec...?
+  context 'when request\'s username has contain invalid format' do
+    let(:base_params) do
+      {
+        employee: {
+          fullname: 'Full Name',
+          username: 'username',
+          password: 'pa55word',
+          password_confirmation: 'pa55word'
+        }
+      }
+    end
+
+    context 'which includes non-ASCII code' do
+      before { post employees_path, params: base_params.merge(employee: { username: 'ユーザー名' }) }
+
+      it { should render_template('new') }
+    end
+
+    context 'which is too short (length < 4)' do
+      before { post employees_path, params: base_params.merge(employee: { username: 'usr' }) }
+
+      it { should render_template('new') }
+    end
+
+    context 'which is too long (length > 64)' do
+      before { post employees_path, params: base_params.merge(employee: { username: 'user' * 17 }) }
+
+      it { should render_template('new') }
+    end
+  end
+end
