+++ spec/requests/employees/edit_spec.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe 'Employees GET /employees/:id/edit', type: :request, js: true do
+  # WARNING: it rarely generates duplicated username by Faker::Internet.username
+  let!(:administor) { create(:administor) }
+  let!(:employees) { create_list(:employee, 5) }
+  let(:employee) { employees.first }
+  let(:employee_id) { employee.id }
+
+  # all actions are requied logged in
+  before { post login_path, params: { username: administor.username, password: administor.password } }
+
+  context 'when employee exsits' do
+    before { get edit_employee_path(employee_id) }
+
+    it { should render_template('edit') }
+
+    it 'can show specified employee' do
+      expect(employee).to eq(assigns[:employee])
+    end
+  end
+
+  context 'when employee does not exist' do
+    before { get edit_employee_path(0) }
+
+    it { should redirect_to(employees_path) }
+  end
+end
