+++ spec/requests/inspections/edit_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe 'Inspections GET /inspections/:id/edit', type: :request, js: true do
+  pending 'Use /ajax/inspections/:id/edit to GET modal'
+end
