+++ spec/requests/orders/edit_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.describe 'Orders GET /orders/:id/edit', type: :request, js: true do
+  pending 'Use /ajax/orders/:id/edit to GET modal'
+end
